+++ src/converter.rs
@@ -1,6 +1,6 @@
 use crate::{Command, Settings};
 use evalexpr::{context_map, eval_float_with_context, DefaultNumericTypes, HashMapContext};
-use gladius_shared::{error::SlicerErrors, types::RetractionType};
+use gladius_shared::{error::SlicerErrors, settings::SettingsPrint, types::RetractionType};
 use std::io::{BufWriter, Write};
 
 pub fn convert(
@@ -12,6 +12,18 @@ pub fn convert(
     let mut layer_count = 0;
     let mut current_object = None;
     let mut write_buf = BufWriter::new(write);
+
+    //output the settings to the gcode file
+
+    for line in settings.to_strings() {
+        writeln!(
+            //lending ; to make comment
+            write_buf,
+            "; {}",
+            line
+        )
+        .map_err(|_| SlicerErrors::FileWriteError)?;
+    }
 
     let start = convert_instructions(
         &settings.starting_instructions,