+++ src/main.rs
@@ -5,6 +5,7 @@
 use clap::Parser;
 use gladius_shared::loader::{Loader, STLLoader, ThreeMFLoader};
 use gladius_shared::types::*;
+use input::load_settings;
 
 use crate::plotter::convert_objects_into_moves;
 use crate::tower::{create_towers, TriangleTower, TriangleTowerIterator};
@@ -22,7 +23,6 @@ use crate::bounds_checking::{check_model_bounds, check_moves_bounds};
 use crate::calculation::calculate_values;
 use crate::command_pass::{CommandPass, OptimizePass, SlowDownLayerPass};
 use crate::converter::convert;
-use crate::input::files_input;
 use crate::plotter::polygon_operations::PolygonOperations;
 use crate::slice_pass::*;
 use crate::slicing::slice;
@@ -49,13 +49,22 @@ mod optimizer;
 mod plotter;
 mod slice_pass;
 mod slicing;
+mod test;
 mod tower;
 mod utils;
 
 #[derive(Parser)]
 #[command(author, version, about, long_about = None)]
+#[clap(group(
+    clap::ArgGroup::new("settings-group")
+        .required(true)
+        .args(&["settings_file_path", "settings_json"]),
+))]
 struct Args {
-    #[arg(required = true)]
+    #[arg(
+        required = true,
+        help = "The input files and there translations.\nBy default it takes a list of json strings that represents how the models should be loaded and translated.\nSee simple_input for an alterantive command. "
+    )]
     input: Vec<String>,
 
     #[arg(short = 'o', help = "Sets the output dir")]
@@ -65,11 +74,22 @@ struct Args {
     verbose: u8,
 
     #[arg(short = 's', help = "Sets the settings file to use")]
-    settings: Option<String>,
-
+    settings_file_path: Option<String>,
+    #[arg(short = 'S', help = "The contents of a json settings file.")]
+    settings_json: Option<String>,
     #[arg(short = 'm', help = "Use the Message System (useful for interprocess communication)")]
     message: bool,
 
+    #[arg(
+        long = "print_settings",
+        help = "Print the final combined settings out to Stdout and Terminate. Verbose level 4 will print but continue."
+    )]
+    print_settings: bool,
+    #[arg(
+        long = "simple_input",
+        help = "The input should only be a list of files that will be auto translated to the center of the build plate."
+    )]
+    simple_input: bool,
     #[arg(
         short = 'j',
         help = "Sets the number of threads to use in the thread pool (defaults to number of CPUs)"
@@ -112,10 +132,37 @@ fn main() {
     }
 
     display_state_update("Loading Inputs", send_messages);
-    let (models, settings) = handle_err_or_return(
-        files_input(args.settings.as_deref(), Some(args.input)),
+
+    let settings_json = args.settings_json.unwrap_or_else(|| {
+        handle_err_or_return(
+            input::load_settings_json(
+                args.settings_file_path
+                    .as_deref()
+                    .expect("CLAP should handle requring a settings option to be Some"),
+            ),
+            send_messages,
+        )
+    });
+
+    let settings = handle_err_or_return(
+        load_settings(args.settings_file_path.as_deref(), &settings_json),
         send_messages,
     );
+
+    let models = handle_err_or_return(
+        crate::input::load_models(Some(args.input), &settings, args.simple_input),
+        send_messages,
+    );
+    if args.print_settings {
+        for line in gladius_shared::settings::SettingsPrint::to_strings(&settings) {
+            println!("{}", line);
+        }
+        std::process::exit(0);
+    } else if log::log_enabled!(log::Level::Trace) {
+        for line in gladius_shared::settings::SettingsPrint::to_strings(&settings) {
+            log::trace!("{}", line);
+        }
+    }
 
     handle_err_or_return(check_model_bounds(&models, &settings), send_messages);